Move statementsIntersect
