+++ pkg/agents/team5/agent.go
@@ -103,84 +103,6 @@ func (a *CustomAgent5) checkForLeader() {
 	}
 }
 
-func statementsIntersect(op1 messages.Op, value1 int, op2 messages.Op, value2 int) bool {
-	switch op1 {
-	case messages.EQ:
-		switch op2 {
-		case messages.EQ:
-			return value1 == value2
-		case messages.LT:
-			return value1 < value2
-		case messages.LE:
-			return value1 <= value2
-		case messages.GT:
-			return value1 > value2
-		case messages.GE:
-			return value1 >= value2
-		default:
-			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
-			return true
-		}
-
-	case messages.LT:
-		switch op2 {
-		case messages.EQ, messages.GE:
-			return value1 > value2
-		case messages.LT, messages.LE:
-			return true
-		case messages.GT:
-			return value1-1 > value2
-		default:
-			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
-			return true
-		}
-
-	case messages.LE:
-		switch op2 {
-		case messages.EQ, messages.GE:
-			return value1 >= value2
-		case messages.LT, messages.LE:
-			return true
-		case messages.GT:
-			return value1 > value2
-		default:
-			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
-			return true
-		}
-
-	case messages.GT:
-		switch op2 {
-		case messages.EQ, messages.LE:
-			return value1 < value2
-		case messages.LT:
-			return value1 < value2-1
-		case messages.GT, messages.GE:
-			return true
-		default:
-			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
-			return true
-		}
-
-	case messages.GE:
-		switch op2 {
-		case messages.EQ, messages.LE:
-			return value1 <= value2
-		case messages.LT:
-			return value1 < value2
-		case messages.GT, messages.GE:
-			return true
-		default:
-			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
-			return true
-		}
-
-	default:
-		// It should be impossible to get here
-		log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
-		return true
-	}
-}
-
 func (a *CustomAgent5) treatiesCanCoexist(t1 messages.Treaty, t2 messages.Treaty) bool {
 	conditionsIntersect := statementsIntersect(t1.ConditionOp(), t1.ConditionValue(), t2.ConditionOp(), t2.ConditionValue())
