+++ pkg/agents/team5/utils.go
@@ -1,6 +1,10 @@
 package team5
 
-import "github.com/google/uuid"
+import (
+	"github.com/SOMAS2021/SOMAS2021/pkg/messages"
+	"github.com/google/uuid"
+	log "github.com/sirupsen/logrus"
+)
 
 func PercentageHP(a *CustomAgent5) int {
 	return int(float64(a.HP()) / float64(a.HealthInfo().MaxHP) * 100.0)
@@ -23,4 +27,82 @@ func (a *CustomAgent5) memoryIdExists(id uuid.UUID) bool {
 
 func (a *CustomAgent5) ResetSurroundingAgents() {
 	a.surroundingAgents = make(map[int]uuid.UUID)
+}
+
+func statementsIntersect(op1 messages.Op, value1 int, op2 messages.Op, value2 int) bool {
+	switch op1 {
+	case messages.EQ:
+		switch op2 {
+		case messages.EQ:
+			return value1 == value2
+		case messages.LT:
+			return value1 < value2
+		case messages.LE:
+			return value1 <= value2
+		case messages.GT:
+			return value1 > value2
+		case messages.GE:
+			return value1 >= value2
+		default:
+			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
+			return true
+		}
+
+	case messages.LT:
+		switch op2 {
+		case messages.EQ, messages.GE:
+			return value1 > value2
+		case messages.LT, messages.LE:
+			return true
+		case messages.GT:
+			return value1-1 > value2
+		default:
+			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
+			return true
+		}
+
+	case messages.LE:
+		switch op2 {
+		case messages.EQ, messages.GE:
+			return value1 >= value2
+		case messages.LT, messages.LE:
+			return true
+		case messages.GT:
+			return value1 > value2
+		default:
+			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
+			return true
+		}
+
+	case messages.GT:
+		switch op2 {
+		case messages.EQ, messages.LE:
+			return value1 < value2
+		case messages.LT:
+			return value1 < value2-1
+		case messages.GT, messages.GE:
+			return true
+		default:
+			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
+			return true
+		}
+
+	case messages.GE:
+		switch op2 {
+		case messages.EQ, messages.LE:
+			return value1 <= value2
+		case messages.LT:
+			return value1 < value2
+		case messages.GT, messages.GE:
+			return true
+		default:
+			log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
+			return true
+		}
+
+	default:
+		// It should be impossible to get here
+		log.Error("Simulation - team5/agent.go: \t Reached unreachable code in statementsIntersect", op1, value1, op2, value2)
+		return true
+	}
 }
